+++ src/main/php/Target.php
@@ -13,14 +13,15 @@ namespace stubbles\webapp;
  *
  * @api
  */
-interface Processor
+interface Target
 {
     /**
-     * processes the request
+     * resolves the request and returns resource data
      *
      * @param  \stubbles\webapp\Request   $request   current request
      * @param  \stubbles\webapp\Response  $response  response to send
      * @param  \stubbles\webapp\UriPath   $uriPath   information about called uri path
+     * @return  mixed
      */
-    public function process(Request $request, Response $response, UriPath $uriPath);
+    public function resolve(Request $request, Response $response, UriPath $uriPath);
 }