+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
    * changed request interface from `stubbles\input\web\WebRequest` to `stubbles\webapp\Request`
    * renamed `stubbles\webapp\response\Response` to `stubbles\webapp\Response`
+   * replaced `stubbles\webapp\Processor` with `stubbles\webapp\Target`
    * both request and response are not available via injection any more
    * session instance must now be created in `stubbles\webapp\Webapp::createSession()` instead of passing a session creator closure to io bindings
    * moved `stubbles\webapp\ioc\Auth` to `stubbles\webapp\auth\Auth`